services/horizon: rename references in services/horizon to point to services/horizon/internal
